refactor: make 'soliplex.cli' a package
- Factor out 'soliplex.cli.types' to break cycles.

- Factor out 'soliplex.cli.cli_utils (ditto).

- New 'soliplex.cli.serve' module

  Move 'soliplex-cli serve' into it (still invoked as 'soliplex-cli serve').

- New 'soliplex.cli.ollama' module:

  'soliplex-cli ollama pull' now replaces 'soliplex pull-models'
  (w hidden BBB alias).

- New 'soliplex.cli.audit' module:

  'soliplex-cli audit installation' now replaces 'soliplex-cli check-config'
   (w hidden BBB alias).

  'soliplex-cli audit secrets' now replaces 'soliplex-cli list-secrets'
   (w hidden BBB alias).

  'soliplex-cli audit environment' now replaces
  'soliplex-cli list-environment' (w hidden BBB alias).

  'soliplex-cli audit oidc' now replaces
  'soliplex-cli list-oidc-auth-providers' (w hidden BBB alias).

  'soliplex-cli audit rooms' now replaces
  'soliplex-cli list-rooms' (w hidden BBB alias).

  'soliplex-cli audit completions' now replaces
  'soliplex-cli list-completions' (w hidden BBB alias).

  'soliplex-cli audit skills' now replaces
  'soliplex-cli list-skills' (w hidden BBB alias).

- New 'soliplex.cli.admin_users' module:

  'soliplex-cli admin-users list' now replaces
  'soliplex-cli list-admin-users' (w hidden BBB alias).

  'soliplex-cli admin-users clear' now replaces
  'soliplex-cli clear-admin-users' (w hidden BBB alias).

  'soliplex-cli admin-users add' now replaces
  'soliplex-cli add-admin-user' (w hidden BBB alias).

- New 'soliplex.cli.room_authz' module:

  'soliplex-cli room-authz show' now replaces
  'soliplex-cli show-room-authz' (w hidden BBB alias).

  'soliplex-cli room-authz clear' now replaces
  'soliplex-cli clear-room-authz' (w hidden BBB alias).

  'soliplex-cli room-authz add-user' now replaces
  'soliplex-cli add-room-user' (w hidden BBB alias).

- Move miscellaneous commands to bottom of help output:

  'soliplex-cli config'

  'soliplex-cli agui-feature-schemas'
