+++ src/soliplex/cli/admin_users.py
@@ -1,0 +1,108 @@
+from __future__ import annotations
+
+import json
+
+import typer
+
+from soliplex.authz import schema as authz_schema
+from soliplex.cli import cli_util
+from soliplex.cli import types
+
+the_console = cli_util.the_console
+
+
+app = typer.Typer(
+    name="admin-users",
+    help="Manage admin users",
+    no_args_is_help=True,
+)
+
+
+def _dump_admin_users(session):
+    with session:
+        admin_users = [
+            admin_user.email
+            for admin_user in session.query(
+                authz_schema.AdminUser,
+            )
+        ]
+    print(json.dumps({"admin_users": admin_users}))
+
+
+@app.command("list")
+def list_admin_users(
+    ctx: typer.Context,
+    installation_path: types.installation_path_type,
+    skip_ram_db_check: bool = typer.Option(
+        False,
+        "-s",
+        "--skip-ram-db-check",
+        help="Skip check for RAM-based DB",
+    ),
+):
+    """Show admin users defined in the installation's authz database."""
+    the_installation = cli_util.get_installation(installation_path)
+    dburi = the_installation.authorization_dburi_sync
+
+    if not skip_ram_db_check:
+        cli_util._check_ram_dburi(dburi, "admin-users list")
+
+    session = authz_schema.get_session(engine_url=dburi, init_schema=True)
+    _dump_admin_users(session)
+
+
+@app.command("clear")
+def clear_admin_users(
+    ctx: typer.Context,
+    installation_path: types.installation_path_type,
+    skip_ram_db_check: bool = typer.Option(
+        False,
+        "-s",
+        "--skip-ram-db-check",
+        help="Skip check for RAM-based DB",
+    ),
+):
+    """Clear admin users from the installation's authz database."""
+    the_installation = cli_util.get_installation(installation_path)
+    dburi = the_installation.authorization_dburi_sync
+
+    if not skip_ram_db_check:
+        cli_util._check_ram_dburi(dburi, "admin-users clear")
+
+    session = authz_schema.get_session(engine_url=dburi, init_schema=True)
+
+    with session:
+        for admin_user in session.query(authz_schema.AdminUser):
+            session.delete(admin_user)
+        session.commit()
+
+        _dump_admin_users(session)
+
+
+@app.command("add")
+def add_admin_user(
+    ctx: typer.Context,
+    installation_path: types.installation_path_type,
+    admin_user_email: str,
+    skip_ram_db_check: bool = typer.Option(
+        False,
+        "-s",
+        "--skip-ram-db-check",
+        help="Skip check for RAM-based DB",
+    ),
+):
+    """Add an admin user to the installation's authz database."""
+    the_installation = cli_util.get_installation(installation_path)
+    dburi = the_installation.authorization_dburi_sync
+
+    if not skip_ram_db_check:
+        cli_util._check_ram_dburi(dburi, "admin-users add")
+
+    session = authz_schema.get_session(engine_url=dburi, init_schema=True)
+
+    with session:
+        admin_user = authz_schema.AdminUser(email=admin_user_email)
+        session.add(admin_user)
+        session.commit()
+
+        _dump_admin_users(session)
